MOved BufferView from include/vsg/vk into incldue/vsg/state
