+++ include/vsg/commands/BindVertexBuffers.h
@@ -15,6 +15,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/commands/Command.h>
 #include <vsg/state/Descriptor.h>
 #include <vsg/vk/Buffer.h>
+#include <vsg/vk/BufferData.h>
 #include <vsg/vk/vk_buffer.h>
 
 namespace vsg
