+++ include/vsg/state/Descriptor.h
@@ -12,8 +12,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 </editor-fold> */
 
-#include <vsg/vk/BufferData.h>
-#include <vsg/vk/BufferView.h>
+#include <vsg/core/Inherit.h>
 
 namespace vsg
 {
