+++ include/vsg/state/Image.h
@@ -17,6 +17,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 namespace vsg
 {
+    // forward declare
     class Context;
 
     class VSG_DECLSPEC Image : public Inherit<Object, Image>
